+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Changed
 - tauri: macOS: webxdc: Remove the nowhere-proxy to support pre-14 macOS. #5202
 - reword 'Save As' to 'Export Attachment' to have a clearer cut to 'Save' #5245
+- use rpc.getWebxdcInfo instead of message.webxdcInfo #5227
 
 ### Fixed
 - tauri: remember webxdc app windows' position and size between app re-launches